pyproject.toml
[project]
name = "canvas"
version = "0.1.0"
description = "Home Assistant integration for Canvas LMS"
readme = "README.md"
authors = [
    { name = "Stephen Vanterpool", email = "StephenVanterpool@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "async-timeout>=5.0.1",
    "homeassistant>=2024.12.5",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aresponses>=3.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
